requirements.txt



docx2txt
google-ai-generativelanguage
google-api-core

jsonschema-specifications
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters

narwhals
nest-asyncio
numpy
openai
orjson
packaging
pandas


python-dotenv


streamlit

tqdm
tzdata
urllib3
watchdog
yarl
zstandard
gradio
fastapi
uvicorn
pypdf
bs4
faiss-cpu